bg-remove-service/pyproject.toml
[project]
name = "ai-service"
version = "0.2.0"
description = "AI Service for background removal and clothing attribute extraction"
authors = [
    {name = "khuongduy354",email = "42113313+khuongduy354@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "transformers (>=4.57.6,<5.0.0)",
    "torch (>=2.9.1,<3.0.0)",
    "torchvision (>=0.24.1,<0.25.0)",
    "scikit-image (>=0.26.0,<0.27.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.1,<0.35.0)",
    "python-multipart (>=0.0.20,<0.1.0)",
    "pillow (>=11.2.0,<12.0.0)",
    "google-genai (>=1.0.0,<2.0.0)",
    "pydantic (>=2.0.0,<3.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "bg_remove_service", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
